refactor: 패키지 분류 및 미사용 클래스 삭제
