+++ src/main/java/racingcar/domain/CarFactory.java
@@ -1,7 +1,5 @@
-package racingcar.service;
+package racingcar.domain;
 
-
-import racingcar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
