+++ src/main/java/racingcar/dto/RequestDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 public class RequestDto {
     String names;
