+++ src/main/java/racingcar/dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 import racingcar.domain.Car;
